context apiの実装
